websocket/requirements.txt
websockets==10.3
websocket-client==1.4.1
redis==4.3.4
google-cloud-vision==3.1.3
google-cloud-texttospeech==2.12.3
jsonpickle===2.2.0